Create new module for error datatypes
